delet console.info + deplace type in index.d.ts
