+++ client/src/types/index.d.ts
@@ -10,3 +10,24 @@ interface CompanyData {
   website: string;
   siret: string;
 }
+
+type OfferData = {
+  title: string;
+  city: string;
+  logo: string;
+  background: string;
+  description: string;
+  date: string;
+  salary: number;
+  skills: string;
+  requirements: string;
+  remote: string;
+  company_id: number;
+  contract_id: number;
+};
+
+interface OfferFormProps {
+  children: ReactNode;
+  defaultValue: OfferData;
+  onSubmit: (offer: OfferData) => void;
+}
